+++ packages/stargate/src/modules/gov/messages.ts
@@ -1,10 +1,11 @@
 import { EncodeObject, GeneratedType } from "@cosmjs/proto-signing";
-import { MsgDeposit, MsgSubmitProposal, MsgVote } from "cosmjs-types/cosmos/gov/v1beta1/tx";
+import { MsgDeposit, MsgSubmitProposal, MsgVote, MsgVoteWeighted } from "cosmjs-types/cosmos/gov/v1beta1/tx";
 
 export const govTypes: ReadonlyArray<[string, GeneratedType]> = [
   ["/cosmos.gov.v1beta1.MsgDeposit", MsgDeposit],
   ["/cosmos.gov.v1beta1.MsgSubmitProposal", MsgSubmitProposal],
   ["/cosmos.gov.v1beta1.MsgVote", MsgVote],
+  ["/cosmos.gov.v1beta1.MsgVoteWeighted", MsgVoteWeighted],
 ];
 
 export interface MsgDepositEncodeObject extends EncodeObject {
@@ -34,4 +35,13 @@ export interface MsgVoteEncodeObject extends EncodeObject {
 
 export function isMsgVoteEncodeObject(object: EncodeObject): object is MsgVoteEncodeObject {
   return (object as MsgVoteEncodeObject).typeUrl === "/cosmos.gov.v1beta1.MsgVote";
+}
+
+export interface MsgVoteWeightedEncodeObject extends EncodeObject {
+  readonly typeUrl: "/cosmos.gov.v1beta1.MsgVoteWeighted";
+  readonly value: Partial<MsgVoteWeighted>;
+}
+
+export function isMsgVoteWeightedEncodeObject(object: EncodeObject): object is MsgVoteWeightedEncodeObject {
+  return (object as MsgVoteWeightedEncodeObject).typeUrl === "/cosmos.gov.v1beta1.MsgVoteWeighted";
 }